Matcher cleanup and readibility
