--- src/test/java/org/cloudfoundry/community/servicebroker/model/fixture/AsyncArgumentMatcher.java
@@ -1,18 +1,0 @@
-package org.cloudfoundry.community.servicebroker.model.fixture;
-
-import org.cloudfoundry.community.servicebroker.model.*;
-import org.mockito.ArgumentMatcher;
-
-public class AsyncArgumentMatcher extends ArgumentMatcher<ServiceInstanceRequest> {
-	
-	static AsyncArgumentMatcher isAsyncRequest() { 
-		return new AsyncArgumentMatcher(); 
-	}
-	
-	@Override
-	public boolean matches(Object argument) {
-        return true == ((ServiceInstanceRequest) argument)
-                .hasAsyncClient();
-	}
-
-}
