+++ src/test/java/org/cloudfoundry/community/servicebroker/model/fixture/ServiceInstanceControllerIntegrationTest.java
@@ -1,13 +1,12 @@
 package org.cloudfoundry.community.servicebroker.model.fixture;
 
+import static org.cloudfoundry.community.servicebroker.model.matchers.AsyncArgumentMatcher.anyAsyncRequest;
+import static org.cloudfoundry.community.servicebroker.model.matchers.SyncArgumentMatcher.anySyncRequest;
 import static org.hamcrest.Matchers.*;
 import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.argThat;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static org.cloudfoundry.community.servicebroker.model.fixture.SyncArgumentMatcher.isSyncRequest;
-import static org.cloudfoundry.community.servicebroker.model.fixture.AsyncArgumentMatcher.isAsyncRequest;
 
 import org.cloudfoundry.community.servicebroker.controller.ServiceInstanceController;
 import org.cloudfoundry.community.servicebroker.exception.*;
@@ -325,7 +324,7 @@ public class ServiceInstanceControllerIntegrationTest {
 	    ServiceInstance instance = ServiceInstanceFixture.getServiceInstance();
 		
 	    when(serviceInstanceService.createServiceInstance(
-	    		(CreateServiceInstanceRequest) argThat(isSyncRequest()))).thenReturn(instance);
+	    		(CreateServiceInstanceRequest) anySyncRequest())).thenReturn(instance);
 	    
 		when(catalogService.getServiceDefinition(any(String.class))).thenReturn(ServiceFixture.getService());
 			    
@@ -442,7 +441,7 @@ public class ServiceInstanceControllerIntegrationTest {
 	    	.withLastOperation(new ServiceInstanceLastOperation("doin stuff", OperationState.IN_PROGRESS));
 	    
 	    when(serviceInstanceService.deleteServiceInstance(
-	    		(DeleteServiceInstanceRequest) argThat(isAsyncRequest()))).thenReturn(instance);
+	    		(DeleteServiceInstanceRequest) anyAsyncRequest())).thenReturn(instance);
 	    
 	    String url = ServiceInstanceController.BASE_PATH + "/" + instance.getServiceInstanceId() 
 	    		+ "?service_id=" + instance.getServiceDefinitionId()
